ansible/files/backend/requirements.txt
flask==2.3.3
psycopg2-binary==2.9.9
redis==5.0.1
prometheus-flask-exporter==0.23.0